Fix no-service-found issue when `apollo service:check`
When running `npx apollo service:check`
https://www.apollographql.com/docs/tutorial/production.html#publish,
It will throw the no-service-found as follows:
```
  ✔ Loading Apollo Project
  ✖ Checking service for changes
    → No service found to link to Engine
Error: No service found to link to Engine
    at Task.task (~/.nvm/versions/node/v8.0.0/lib/node_modules/apollo/lib/commands/service/check.js:16:31)
    at Promise.resolve.then.then.skipped (~/.nvm/versions/node/v8.0.0/lib/node_modules/apollo/node_modules/listr/lib/task.js:167:30)
    at <anonymous>
```
The reason is the apollo.config.js is missed to config the service name.

The side effect of this change is to force the user to name their
service as `space-explorer`, which is later used in the client,
so I think it's better to force the user to use the name to avoid
confusing and unnecessary error.